Move migrations to core bundle
